Java code refactoring
