+++ app/src/common/shared/org/mozilla/vrbrowser/VRBrowserActivity.java
@@ -31,17 +31,20 @@ import org.mozilla.geckoview.GeckoRuntime;
 import org.mozilla.geckoview.GeckoSession;
 import org.mozilla.vrbrowser.audio.AudioEngine;
 import org.mozilla.vrbrowser.audio.VRAudioTheme;
+import org.mozilla.vrbrowser.input.MotionEventGenerator;
 import org.mozilla.vrbrowser.search.SearchEngine;
 import org.mozilla.vrbrowser.telemetry.TelemetryWrapper;
-import org.mozilla.vrbrowser.ui.BrowserWidget;
-import org.mozilla.vrbrowser.ui.CrashDialogWidget;
-import org.mozilla.vrbrowser.ui.KeyboardWidget;
-import org.mozilla.vrbrowser.ui.NavigationBarWidget;
+import org.mozilla.vrbrowser.ui.widgets.BrowserWidget;
+import org.mozilla.vrbrowser.ui.widgets.CrashDialogWidget;
+import org.mozilla.vrbrowser.ui.widgets.KeyboardWidget;
+import org.mozilla.vrbrowser.ui.widgets.NavigationBarWidget;
 import org.mozilla.vrbrowser.ui.OffscreenDisplay;
-import org.mozilla.vrbrowser.ui.RootWidget;
-import org.mozilla.vrbrowser.ui.TopBarWidget;
-import org.mozilla.vrbrowser.ui.TrayWidget;
-import org.mozilla.vrbrowser.ui.UIWidget;
+import org.mozilla.vrbrowser.ui.widgets.RootWidget;
+import org.mozilla.vrbrowser.ui.widgets.TopBarWidget;
+import org.mozilla.vrbrowser.ui.widgets.TrayWidget;
+import org.mozilla.vrbrowser.ui.widgets.Widget;
+import org.mozilla.vrbrowser.ui.widgets.WidgetManagerDelegate;
+import org.mozilla.vrbrowser.ui.widgets.WidgetPlacement;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
