+++ app/src/common/shared/org/mozilla/vrbrowser/ui/views/CustomUIButton.java
@@ -1,4 +1,4 @@
-package org.mozilla.vrbrowser.ui;
+package org.mozilla.vrbrowser.ui.views;
 
 import android.graphics.drawable.Drawable;
