+++ app/src/common/shared/org/mozilla/vrbrowser/ui/views/NavigationURLBar.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.vrbrowser.ui;
+package org.mozilla.vrbrowser.ui.views;
 
 import android.content.Context;
 import android.content.res.Resources;
@@ -206,7 +206,7 @@ public class NavigationURLBar extends FrameLayout {
         mURL.setPadding(leftPadding, mURL.getPaddingTop(), mURL.getPaddingRight(), mURL.getPaddingBottom());
     }
 
-    protected void handleURLEdit(String text) {
+    public  void handleURLEdit(String text) {
         text = text.trim();
         URI uri = null;
         try {
