+++ app/src/common/shared/org/mozilla/vrbrowser/ui/widgets/RestartDialogWidget.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.vrbrowser.ui;
+package org.mozilla.vrbrowser.ui.widgets;
 
 import android.app.AlarmManager;
 import android.app.PendingIntent;
@@ -14,8 +14,8 @@ import android.view.View;
 import android.widget.TextView;
 import org.mozilla.vrbrowser.R;
 import org.mozilla.vrbrowser.VRBrowserActivity;
-import org.mozilla.vrbrowser.WidgetPlacement;
 import org.mozilla.vrbrowser.audio.AudioEngine;
+import org.mozilla.vrbrowser.ui.views.UIButton;
 
 public class RestartDialogWidget extends UIWidget {
     private static final String LOGTAG = "VRB";
