+++ app/src/common/shared/org/mozilla/vrbrowser/ui/widgets/SettingsWidget.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.vrbrowser.ui;
+package org.mozilla.vrbrowser.ui.widgets;
 
 import android.content.Context;
 import android.content.pm.PackageInfo;
@@ -21,6 +21,7 @@ import android.widget.TextView;
 import org.mozilla.geckoview.GeckoSession;
 import org.mozilla.vrbrowser.*;
 import org.mozilla.vrbrowser.audio.AudioEngine;
+import org.mozilla.vrbrowser.ui.views.SettingsButton;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
