+++ app/src/common/shared/org/mozilla/vrbrowser/ui/widgets/Widget.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.vrbrowser;
+package org.mozilla.vrbrowser.ui.widgets;
 
 import android.graphics.SurfaceTexture;
 import android.view.MotionEvent;
