+++ app/src/main/res/layout/close_button.xml
@@ -4,7 +4,7 @@
               android:layout_width="match_parent"
               android:layout_height="match_parent"
               android:orientation="horizontal">
-<org.mozilla.vrbrowser.ui.UIButton
+<org.mozilla.vrbrowser.ui.views.UIButton
         android:id="@+id/closeButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
