+++ app/src/main/res/layout/setting_edit.xml
@@ -45,7 +45,7 @@
                     android:visibility="visible"
                     tools:text="Text 1"/>
 
-                <org.mozilla.vrbrowser.ui.DeveloperOptionsEditText
+                <org.mozilla.vrbrowser.ui.views.DeveloperOptionsEditText
                     android:id="@+id/edit1"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -75,7 +75,7 @@
                     android:visibility="visible"
                     tools:text="Text 2" />
 
-                <org.mozilla.vrbrowser.ui.DeveloperOptionsEditText
+                <org.mozilla.vrbrowser.ui.views.DeveloperOptionsEditText
                     android:id="@+id/edit2"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
