+++ app/src/main/res/layout/tray.xml
@@ -11,19 +11,19 @@
             android:orientation="horizontal"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content">
-        <org.mozilla.vrbrowser.ui.UIButton
+        <org.mozilla.vrbrowser.ui.views.UIButton
                 android:id="@+id/helpButton"
                 style="?attr/trayButtonStyle"
                 app:privateModeBackground="@drawable/tray_button_background"
                 android:src="@drawable/ic_icon_tray_help_v2"
                 android:rotation="-15" android:translationX="5dp" android:translationY="0dp"
                 android:translationZ="0dp"/>
-        <org.mozilla.vrbrowser.ui.UIButton
+        <org.mozilla.vrbrowser.ui.views.UIButton
                 android:id="@+id/privateButton"
                 style="?attr/trayButtonStyle"
                 android:src="@drawable/ic_tray_private"
                 android:layout_marginBottom="5dp"/>
-        <org.mozilla.vrbrowser.ui.UIButton
+        <org.mozilla.vrbrowser.ui.views.UIButton
                 android:id="@+id/settingsButton"
                 style="?attr/trayButtonStyle"
                 app:privateModeBackground="@drawable/tray_button_background"
